+++ src/redux/actions/coursesAction.js
@@ -5,6 +5,9 @@ const getCoursesAction = () => dispatch => {
     return coursesRequests.getAllCourses().then(res => dispatch({
         type: types.courses.FETCH_COURSES_SUCCESS,
         payload: res.data
+    })).catch(err => dispatch({
+        type: types.courses.FETCH_COURSES_FAILED,
+        payload: err.error
     }));
 }
